fix adding new titles
when a new title was added, the current title was used for comparison,
not the last title in the playlist. So it could come to repetitions of
artists.

add mpplayer.html

mpplayer is a tiny HTML/js combo that uses the HTML audio tag to play
the current title. This has no further controls, no fading or anything
else and should not be taken seriously.