prepare better search
searching now returns found titles, artists and albums and forwards the
results to the client. In this effort some bugs in the json interface
were fixed, mainly CBV/CBR misunderstandings.

Also the command structure was changed. The new/modified commands are:
- mpc_search <keyword> - triggers the search with the range set in the
command and the keyword in the argument
- mpc_insert <cnt> - plays the cnt's result next. If cnt is -1 all
results will be used, range set on the command decides if it's titles,
albums or artists
- mpc_append <cnt> - like above just at the end of the playlist

Search is limited to 100 results - and even this is debatable IMHO.

A new json type MPCOMM_RESULT has been introduced to mark a reply to
contain searchresults. Also titles now contain their key.

Of course the protocol version has been bumped too.