rework search
to have sensible searchresults and a way to handle them some changes had
to be made to the range definitions. The search ranges are now bitflags
so it is possible to search for artists and albums at the same time.

The functions to add a title to a playlist now have a flag to control if
markings shall be taken into account. This is needed to adding a title
to a result playlist does not mark it as played.

The web interface had to be adapted to the new ranges, also the choice
of search ranges was limited.

Searching for albums also saves the artist so when two artists have an
album of the same name these can be told apart.

During the search rework the [talg][*=] notation was dropped. This will
probably happen for FAV/DNP as well (See #18 for followups)

This change does not yet implement any client to use the search results
and may break cmixplay and gmixplay.