improve UI handling
the search function is now available und usable in the web client.

In the flow of this some quirks have been fixed. The server no longer sends full status updates all the time. This also fixes the broken scroll of long title lines.

Layout of playlist and search results needs to be tweaked but that is more CSS than programming.

The JSON support now sends UTF and does not filter it out. Also the music scan now adds special chars and does no longer filter them out quietly. This it's recommended to remove the DB and rescan.