reduce media wear during play
so far every time a title finished playing the playcount was increased
and the title was written back into the db immediately. This may cause
unwanted wear on the media the database is stored on. Now the database
is just marked dirty and on a normal termination the database is written
back with the updated results. The same goes for cleaning and adding new
titles.

Also single bit flags in the config have been put into a bitfield.