do not increase count on inserted titles
after searching and playing a title, it has been marked as counted and
the playcount was increased, also it would have been marked as skipped
if the next title was forced. Now searchresults are handled as out of
order titles.

fix stepping back the playlist for more than one title.

add a warning if more than 100 Titles were found in a search.

move from strlcat/strlcpy to own strtcat/strtcpy as the needed
functionality differs a little from the strl* versions.