fix bug in strtcat
the final length calculation was wrong so strings may get cut up in the
middle when setting the guaranteed null byte.

fade flag is now a single bit. This may change again when playerless
server comes into play.

During scan it may have happened that a title with a zero length artist
was inserted into the database which broke the player. This means the
database needs a full rescan!

the mpg123 processes do not interpret mp3v2 tags any longer