remove stand alone clients
since the standalone clients are no longer tested and barely even used I 
decided to remove them once again.

Maybe there will be some kind of client library with a simple client 
example in the future but even that is questionable as it doesn't really 
fit the main use case. So far the web interface works best for both 
local (use as web-app) and remote (in browser).