add titles to current playlist
new titles were only added to the database on the disk. So when the 
player quit, the new titles were removed when the current titles were 
written. Also it was not possible to search and play the new titles. 

start comm server first so a blocked socket keeps the other threads from 
starting. This should lower the probability for zombified threads and 
mpg123 subprocesses.

raise debug level for success messages in tag parsing.

fix clean up of playlist conrnercase. When every entry in the list was 
removed plCheck() crashed.

try to adapt client window according to stream or normal paly 
(experimental!)