implement playlist notification
handling playlist updates in the UI included a lot of guesswork and 
synchronization issues. This patch adds a notifyChange callback that 
notifies every listener that the playlist has changed. In the HTTP 
client this means a full update is necessary.

So far this is used merely internally but shall become the framework for 
all display clients.

Do not wipe found titles after adding them. The player will crash if the 
client wants to add them a second time.

Tweak warning level on some messages from always to on debug.