Prepare favplay
favplay will be a mode where only favourites will be played. Like a 
playlist defined with the FAV/DNP syntax. Search and add shall be 
handled dependant on play/edit mode.

Remove flags from config and join them into new mpmode field. Add 
control macros for easier handling.

change applyFAVlist to accept an 'exclusive' parameter flag. If this is 
set, all titles will be marked DNP first, then the favlist is applied.

TODO: There is no way to use this yet and the UI is unaware of the 
protocol change.