Remove mixplay playlists
creating and supporting own playlists has proven to not bring the 
expected effect. Instead the player now supports a 'Favplay' mode, in 
which only favourites are played.

Also the search now supports a 'Fav' mode in which the results can be 
marked as favourites and not added to the playlist.

If the player is in Favplay mode the PlaySearch returns only favourites, 
the FavSearch returns only DNPs.