make search request synchronous
synchronizing a search across threads, clients and connections is a real 
hassle. Now the search call is synchronous, so the answer is forced to 
be in the call-reply from the server.

To prepare for the changed logic in the client, the updateUI function 
has been split into special update functions to keep the code a little 
more maintainable. Although it may be time to split the javascript too.