add client API
a proof of concept implemetation of a client API that can send commands 
and get the current title. The mixplay-hid is a demo application that 
allows controlling mixplayd with the keyboard.

This may become an option for mixplayd so it stays in the foreground and 
allows easier support for flirc

While trying to trim down the needed objects, some unneeded dependencies 
have been resolved. So this is more a code cleanup than adding any real 
functionality.