basic HID support in mixplayd
when started with -d mixplayd will react to keypresses in the current 
terminal. So far these are just the most commonly used functions. These 
will be configurable, so special keypads and flirc can be supported more 
easily.

the debug level to transmit playcount to the client has been raised to 
2.