fix more deadlocks
the flow control in the ePaper relied on config->command and not 
config-status. This means that mpc_quit was only recogniced if it was 
the current command. However a tiny delay may have set that to mpc_idle 
again while the status is stuck on quit. Shutting down now should show 
the centered heart again.