fix dnp/fav on artist and album
clicking the dnp/fav popups for album and artist had no effect.

Change cursor properly when hovering over UI elements.

Do not propagate 'space' when used as pause/play to avoid accidental 
button events. In this line the focus frame was removed for buttons.

Add a little x to default popups as a hint for closing.

Get rid of message tab and put messages into the tools tab.