add DNP to search results
it is now possible to search for titles/albums/artists and mark them as 
DNP in the search tab.

The plcheck function was prone to remove all titles for fun.

After searching for doublets, the list is applied to the current 
database in use.

Increase maximum number ofg search results to 256

Remove doublet button from UI. The function can be called with shift-d 
and giving the password instead.

And some minor tweaks