Fix fail() on UI
the request to retry did retry on cancel and still try to update in any
case.

offer to replay played titles. Titles in the history can be replayed now.
It may be worthwhile to pull up titles to be played too but that would
need a popup with three choices or implementation of drag and drop on
the playlist.