support passive event listeners
chrome always complains about wheel and touch events not being marked as 
passive, not this has eben fixed.

the number of streams and profiles has been removed from the json update 
as it just duplicates the length of the arrays.

Search results has been limited to 100 titles.

page sizes and layout have been tweaked to fit mobile and desktop better