do not start if socket is blocked
starting the http server as a new thread could cause a bind() error 
while the rest of the system is already up and runing, causing an 
abnormal exit. On streamplay this has the chance that the mpg123 process 
stays up in the background while the player terminated.

Further tweaks on the layout