scale buttonbars less than content
the buttonbars now scale with 2/3 the rate of the content. So a fullscreen player no longer looks too weird with huge buttons but still has enough magnification to allow proper touch.

minor changes in streamplay text when no info has been sent yet.

when removing a stream, the profile counter was decreased too.

the scrollwheel now works globally as scrolling the viewport no longer makes proper sense with a absolute positioning of the bottom toolbar.