rework favourites handling
favourites may no longer increase playcount above the normal titles but 
have an extra counter to control extra plays. During favplay only the 
favcounter is used, not affecting the playcount in any way.

The dbinfo function now smoothes playcount. If the spread of playcount 
becomes too wide, this tries to collapse titles in a playcountgroup that 
is smaller than 0.5% of all titles with the next higher group. This 
reduces the maximum playcount and removes unicorns that mess up mixing 
on other profiles.