remove 'Done.' after search
after a search the message log was spammend with a 'Done.' message which 
should just end the progress. Now the progress is ended quietly. This 
also allowder to remove a special state from the addMessage verbosity.

Added -r flag on startup to retry if the socket is blocked.

reduce favpcount too when redusing playcount on leveling playcount.