clean up play/skipcount handling
playcount and skipcount are no longer handled in two places. When the 
current title ends, playcount gets increased and skipcount will be 
increased when mpc_next was sent before. Removing a title from the 
playlist now also counts as a play and skip.

running mixplayd with a single -d no longer sends playcount to the 
client.

replace some malloc()s with strdup()s