add more communication to mpclient
the mixplay client functions now allow to request a status update which 
is returned as a jsonObject and can be handled with the json* functions. 
The mixplay-hid client has been changed to fetch a real status update 
and not just the proper title and show how to access data in the 
returned object. Also some more detailed error handling/messaging has 
been added.

jsonObjects can now always be fully freed, no more distinction between 
key and value.

jsonPeek() can be used to make sure a json key is set.