make popups mutually exclusive
it was possible to have more than one popup open at a time, which did 
not make a lot of sense. Additionally teh active popup line will be 
magnified to make it easier to hit the popup selections on a 
touchscreen.

#39 The UI now loads and runs on Microsoft Edge too. The title info on 
the main page will neither scroll nor be centered though.

The doublet list is now hidden when a stream is played.

add some padding to the buttons