streamline status request
a status request always registers an updatehandler and sends the full 
playlist on first call. This may not be desired by clients like the 
screensaver that sends regular requests but is not interested in full 
updates. So requesting mpctrl/status?## explicitly, the request will be 
handled as a one shot. No update handler and the connection will be 
closed off after sending the reply.

If the request comes through a registered update channel, nothing will 
change.

The client code will treat calls with fd==-1 as one-shot requsts by 
default.