send searchresults to the proper client
since the searchstate was kept in the config, every client could fetch 
the current searchresults and, even worse, try to unlock the client. Now 
the clienthandler has a marker for an active search.

Change MPCOMM_FULLSTAT to MPCOMM_TITLES as FULLSTAT is misleading and 
implying that all data is transferred.

Yet another fix for the skipping first title on change to profile issue.