Add global playlist lock
when changing profile/playlist, the current titles and playlist will be 
wiped. If an update client tries to read the current titles while the 
list is wiped, the program will segfault. Thus a global lock has been 
introduced, that makes sure that an update client will not read on a 
vanishing list.

Additional locks may be neccesary as setProfile() is an extra thread.

Do not put the client in debug mode if the server was started with a 
single -d