Switch to clock on pause
since none of the pause screens were really nice, the default is now to 
switch to the clock display after ten seconds.

Also every touch/click resets the counter, so the screen will not switch 
to the watch while there is active interaction.

Alas, mixplays-scr will remain unfazed by that..