recover from network change
if the network changes while playing a stream (e.g.: a DSL reconnect) 
mpg123 may get stuck and force a hard restart of the whole player. Now 
this case is tried to get recognized and the mpg123 processes will get 
restarted in such a case.

Also the HID runs as an own thread to allow better control of the main 
thread.