move titles in playlist
there is not an item in the title pop-up that allows to move the title 
in the playlist. On the desktop titles can still be dragged with the 
mouse on a touchscreen this doesn't work as nice as the browser tends to 
mis-interpret a drag as a swipe gesture.

keep player from locking up when an argument is not free'd

make title move insert the title before the selected target, replacing 
it, instead of inserting it after.

remove some magic integers

straighten locking for file sends