prioritize DNP/FAV
before the rule was that DNP was beating FAV in any case. Now this has 
been prioritized to easily realize use-cases like
a bad Album with one good title.
Before that meant that each title but the good one needed to be marked 
DNP. Now the album can be marked DNP and one title as FAV. The explicit 
favourite will beat the generic DNP and the title will be played.

Doublets are un-beatable DNPs