clean up command
the 'edit' sommand was used to set the search behaviour and that was set 
in the fpcurrent flag of the config. This has been cleaned up, also it 
is now possible to search in the DNPs even when not in favplay mode.

The socket handling has been revamped to better recover on reload based 
client/socket mix-ups.