harden stream player
on certain network failure conditions, the player may get caught in a 
restart loop. MAde sure that the player does not try to re-connect twice 
to an unreachable channel.

make sure to unlock all mutexes on restart so other threads will 
continue gracefully.

the client may send one-shot requests on the update channel, this closed 
the update channel and froze updates on the client. Now the channel is 
kept open until the reason is identified properly.

add info to error message on failing dowrite, so the offset is known

print PID on debug so one can easily attach with GDB