turn channels into profiles
the profile_t was extended to additionally hold a default volume and a
stream name. Now streams are held in the same structure that profiles
are. This simplifies a number of steps and enables to save current
volumes so differences between channels and local plays no longer need
manual correction.

In this process yet another chance for memory corruption has been fixed
when removing a profile and some string manipulation was streamlined as
strdup is still a thing.