remember last profile
switching from a profile to a stream and back should not cause the 
player to re-apply fav/dnp. So if the player switches to a profile it is 
checked if that profile was the last played profile and in that case 
playing should commence immediately. Otherwise the normal clean and 
dnp/fav application is done.