move syslog to addMessage()
like this only one update handler needs to be registered and even that 
one may undergo another streamlining.

Remembering the last active profile was moved from the config to a 
static local variable in setProfile()

added an activity on changing the profile to avoid confusing messages 
about applying the DNP list.