try to avoid crash in ALSA
when switching to a stream it may happen, that ALSA crashes on the 
snd_mixer_selem_has_playback_switch call, as the argument seems invalid. 
As this is a pain to recreate the code tries to catch the case and just 
do nothing. This may or may not fix the issue unless it's a race 
condition somewhere.