change volume control
on resistive touchscreens the swipe action to control the volume will
somethimes select elements rather than do an actual drag. Now the action
was reverted to tap-control. The client will no longer mute volume but
be able to unmute. Probably this was not the last change for this..

all console.log outputs are now suppressed and only printed if debug
mode is enabled. Just running the server in interactive mode (-d) is
not enough to enable debug on client, this needs at least two -d
options.