lock updates in ui
while it should not happen it is not impossible that an updateUI call 
may trigger while the last call is still active. This change adds a 
simple lock to avoid this.

an mpstop script was added to simply stop a server from the commandline. 
In exchange the 'quit' option was removed from the UI.