Work on stuck player after switching channels
when switching streams it may still happen that the stream is loaded but 
not playing. The player is stuck on idle without allowing the web client 
to start the stream.

tweaked message levels to avoid spam

try to improve startup and channel switching be redefining the meaning 
of mpc_idle, mpc_start and mpc_stop

add mpc_pause state

add displaying state and command in web client with debug enabled

fix linebreaks on mpstop