always run setProfile on init
after a failed connection the player should revert to the last station. 
But since there is already a database, the player just starts, catching 
itself in an endless loop.

Closing the audio connection did not clean up properly, so on restart 
the player may try to access a free()d value.

Streamlined to command handling to remove non-functional exceptions