minify javascript
to reduce application size and startup time, the javascript will be 
minified before adding it to the application.

This also means that the minified javascript will be used on debug runs. 
So after changing the script and reloading the client, the minifier must 
be run like

make build/mixplayd_min.js

The minifier is an additional utility that's built on demand and may 
change over time.