attempt to fix endless loop #45
when skipping titles the musicmanager may not notice a full circle on
the database since the guarding title may be marked as being in the
playlist.

* debug print repeats to identify an upper limit

* write database on ctrl-c

investigate playcount rounds

give more precedence to favourites

rename cleanFavPlay() to cleanTitles()

cleanFavPlay() is misleading as this function cleans up playlist and
titles whenever the profile is changed.