do not try to guess artist spread
deciding how many titles can be played before an artist must repeat was 
guessed by dividing the number of titles by 25. This kind of works for 
large collections but fails on favplay for example. Now the spread is 
computed on changes of the titles, like switching a profile, adding DNPs 
or en-/disabling favplay.

Also the TDARK flag is no longer wiped when the playlist has less than 
10 entries (or the max spread if that is smaller than 10) so filling up 
the initial playlist should speed up some more.