add restart option to mpstop
unset playcount tags before computing artist spread. When setting a DNP 
mark while there are just a few artists with a low playcount it may 
happen that just the few artists with low playcount are considered, 
leading to the dreaded repetition of those artists until the spread is 
computed again.