rework power ogic in UI
the state of the player was reflected on several places in the UI code. 
Now it is derived from the explicit state and no longer guessed on 
commands or playmodes.

Also the logic to handle taps on the screen has been revamped as there 
is a subtle difference between the idle clock and the play clock state.