speed up dnp
when marking the current title as DNP the player may take a while until 
something happens. When pressing the button, the title is marked as DNP, 
removed from the playlist and the playlist is filled up again before 
switching to the next title. Filling up the playlist may take more than 
a second, causing the delay. To avoid this, the order is now: remove 
title(s), check if the current title was removed and if so, play the 
next then fill up the playlist - like on a 'normal' title change.