keep DNP titles that were searched
when checking the playlist, titles marked as DNP are removed. However a 
search may add a title that is marked as DNP and which should not be 
removed. So only test titles that are marked as being added normally 
(MP_INPL)

the same logic is applied to the doublet checking. Titles added through 
search will not count as replayed artists.