fix memory leaks
ending the player after copying a profile lead to a double free(). So the player was run in Valgrind and all reports happening on profile commands were addressed.

In this run also deprecated profile handling code was removed and list handling is done implicitly in the config.c functions and not explicitly in the player.