Lineout volume switch (#70)
* add linestream modifier

commonly streams play louder than local MP3s, so allow volume boosting for profiles.

* Clean up controlVolume()

better support for lineout mode

* change volume when new profile starts

the volume changed when the new profile was selectected. So when changing from a stream to a mix, the stream's volume was already changed before the mix even started playing.

* add linestream config

the lineout volume adaption should not be a constant value, so turn it into a config value.

Update default volume setting for new profiles.