remove search mode
switching the search mode between fav and dnp was not very intuitive.
Now only doublets are  filtered out from the search but all results do have flags attached if they are flagged as dnp or fav. For titles this is trivial as the flags are attched to the mptitle_t. For artists and albums, the fav- and dnplist is checked if the curent entry is explicitly marked. In that case the flag is added and the client can decide how to handle that information.