#63 Upload redesign (#77)
* restructure mpserver

pull the big loop into three parts:
 * fetchRequest()
 * parseRequest()
 * sendReply()

* State cleanup

* add content-length to every reply

some browsers may complain, even on a 204

* auto-add new title

if you upload a new title it will added to the playlist by default

add some new bool's instead of uint32_t's

* avoid truncated first document

* add progress bar for uploads

* add check for existing file in database

use fetch() instead ox xmlhttprequest to post the form. Like this the upload can be cancelled by the server without needing to either still send all the data or trigger retries.

this also means that the server now can send an abort status to cancel the current upload.

upload now has it's own resource path to reduce guesswork.

the newly added title will mark the database as dirty.

* minify broke js promises

* turn errors into alerts

* lock upload to one client

uploading a file could interfere with other clients. It still will show the progress on every client but that has the nice side effect to remove the upload form, so it's fair.

* Clean up locking on data commands

when a command expects replies, the client should be locked, so send a clientid for those requests.

clean up unused states in server and avoid direct send() calls

* fix xmlhttp reply handling

this could potentially improve async commands like search

* remove wrong semicolons

some replies were adding a ';' to the content-length which is not correct.
This lead to chrome not finishing downloads and firefox crashing on Android.

* harden upload for mobile clients

there is still a protocol error hidden, but neither chrome, brave nor firefox want to tell WHAT is wrong with the reply and just gives the equivalent of the yellow motor light on the dashboard.

* Fix comments and content-lengths in replies

* fix alert format issues

* keep HTTP version to 1.0

* fix open() call

Properly calculate filesize from content-length

* fix some coverity issues

mainly create a sendloop() function to take care of the way send() may only send part of the data.

* return albums on last ten

* clean up client locking

* allow more cmds on streamplay

* speed up multiple messages

* remove CC definition in Makefile

code builds with gcc and clang, so let make decide on the default toolchain

* remove search mode

switching the search mode between fav and dnp was not very intuitive.
Now only doublets are  filtered out from the search but all results do have flags attached if they are flagged as dnp or fav. For titles this is trivial as the flags are attched to the mptitle_t. For artists and albums, the fav- and dnplist is checked if the curent entry is explicitly marked. In that case the flag is added and the client can decide how to handle that information.

* color code dnp and fav on searchresults

* color titles too

* switch title and artist in search

in the searchresults the title is usually more interesting than the artist, so swap them in the display.

* fix edge case on last ten

when the last album is a sampler this is not recognized correctly, so do additional test,