Use a threadpool for the connections (#79)
* clean up code

- add assertion in json code to avoid creating loops
- add help in HID code
- clean up screensaver code
- clean up HTTP header formating in server
- code style fixed in musicmanager
- remove unused styles in css

* add connection thread pool

since every connection was creating and destroying a handler thread, this has been moved to a threadpool. Ten threads are ready to handle connections.

* allow stop on start

added a -s option to mixplayd which will keep the
player from immediately starting to play music.
The use-case is for a standalone player that
reboots after a power fail and should probably
not start blasting music.